+++ bsi0007-tilt-sensor/src/BSI0007.cpp
@@ -1,4 +1,4 @@
-#include "TiltSensor.h"
+#include "BSI0007.h"
 
 using namespace wlp::BSI0007;
