+++ bsp008z-pressure-sensor/src/BSP008Z.h
@@ -1,4 +1,4 @@
-#include "PressureSensor.h"
+#include "BSP008Z.h"
 
 using namespace wlp::BSP008Z;
