+++ tmp36-temperature-sensor/src/TMP36.cpp
@@ -1,4 +1,4 @@
-#include "TemperatureSensor.h"
+#include "TMP36.h"
 
 using namespace wlp::TMP36;
